// frontend/src/types/adminSupportResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AdminSupportResponseType } from './adminSupportResponseType';
import type { AdminSupportResponseSeverity } from './adminSupportResponseSeverity';
import type { AdminSupportResponseStatus } from './adminSupportResponseStatus';
import type { FileResponse } from './fileResponse';

export interface AdminSupportResponse {
  id?: number;
  version?: number;
  createdBy?: number;
  updatedBy?: number;
  createdAt?: string;
  updatedAt?: string;
  title?: string;
  description?: string;
  type?: AdminSupportResponseType;
  severity?: AdminSupportResponseSeverity;
  userId?: number;
  userEmail?: string;
  userName?: string;
  status?: AdminSupportResponseStatus;
  reply?: string;
  attachments?: FileResponse[];
  note?: string;
}