// frontend/src/types/adminSupportResponseStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AdminSupportResponseStatus = typeof AdminSupportResponseStatus[keyof typeof AdminSupportResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminSupportResponseStatus = {
  OPEN: 'OPEN',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED',
} as const;